Add repository list
Implemented the repository list for the github profile
page that includes a header with a title which is the
username of the repository owner and a short description.
As well as maximum 4 repositories that displays name of the
repository, a short description, the no of forks and the no
of stars as well as the license and the last updated.